+++ webapp/backend/pkg/models/measurements/smart_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/analogj/scrutiny/webapp/backend/pkg/models/collector"
 	"github.com/analogj/scrutiny/webapp/backend/pkg/models/measurements"
 	"github.com/golang/mock/gomock"
+	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 )
 
@@ -214,7 +215,7 @@ func TestNewSmartFromInfluxDB_ATA(t *testing.T) {
 	}
 
 	//test
-	smart, err := measurements.NewSmartFromInfluxDB(attrs)
+	smart, err := measurements.NewSmartFromInfluxDB(attrs, logrus.New())
 
 	//assert
 	require.NoError(t, err)
@@ -258,7 +259,7 @@ func TestNewSmartFromInfluxDB_NVMe(t *testing.T) {
 	}
 
 	//test
-	smart, err := measurements.NewSmartFromInfluxDB(attrs)
+	smart, err := measurements.NewSmartFromInfluxDB(attrs, logrus.New())
 
 	//assert
 	require.NoError(t, err)
@@ -297,7 +298,7 @@ func TestNewSmartFromInfluxDB_SCSI(t *testing.T) {
 	}
 
 	//test
-	smart, err := measurements.NewSmartFromInfluxDB(attrs)
+	smart, err := measurements.NewSmartFromInfluxDB(attrs, logrus.New())
 
 	//assert
 	require.NoError(t, err)
@@ -348,6 +349,9 @@ func TestFromCollectorSmartInfo(t *testing.T) {
 	//check that temperature was correctly parsed
 	require.Equal(t, int64(163210330144), smartMdl.Attributes["194"].(*measurements.SmartAtaAttribute).RawValue)
 	require.Equal(t, int64(32), smartMdl.Attributes["194"].(*measurements.SmartAtaAttribute).TransformedValue)
+
+	//check that power-on hours was correctly transformed
+	require.Equal(t, int64(1730), smartMdl.Attributes["9"].(*measurements.SmartAtaAttribute).TransformedValue)
 
 	//ensure that Scrutiny warning for a non critical attribute does not set device status to failed.
 	require.Equal(t, pkg.AttributeStatusWarningScrutiny, smartMdl.Attributes["3"].GetStatus())
@@ -740,7 +744,7 @@ func TestNewSmartFromInfluxDB_WithDeviceStatistics(t *testing.T) {
 	}
 
 	//test
-	smart, err := measurements.NewSmartFromInfluxDB(attrs)
+	smart, err := measurements.NewSmartFromInfluxDB(attrs, logrus.New())
 
 	//assert
 	require.NoError(t, err)